Refactor named tuple semantic analysis to a new module
